refactor(balius-runtime): split u5c adapter into mod.rs + convert.rs
The u5c ledger module grew large enough that the ledger-facing surface
(Config, Ledger, wit ↔ utxorpc From impls, chain_utxo_to_wit) was sharing
a file with the u5c → balius_proto conversion layer. Promote
`ledgers/u5c.rs` to a folder with two siblings:

  ledgers/u5c/
    mod.rs        — ledger adapter: Config, Ledger, pattern conversions
    convert.rs    — ConvertError, convert_tx / convert_tx_output / ...,
                    pparams_to_legacy_json, all the BigInt-flattening
                    helpers, and the unit tests that assert ABI compat
                    against utxorpc-spec 0.17 (prost + pbjson)

The previously integration-style tests at
`balius-runtime/tests/u5c_convert.rs` move into a `#[cfg(test)] mod tests`
inside `convert.rs` — they're testing internals of that module, so unit
tests are the right home. Test count unchanged; they now run as part of
the runtime's lib tests.

Call sites in `lib.rs` updated to the new path
(`ledgers::u5c::convert::convert_tx`, `ledgers::u5c::convert::ConvertError`).
No re-exports — the structure is explicit.

Verified: cargo check --workspace clean; 9/9 lib tests pass (8 convert
tests + 1 router test); e2e::faucet_claim still passes; pre-existing
u5c-chainsync::wallet_balance stale-wasm failure unchanged.
